move convolution patch flipping to win...nce.m
Previously, if C2.m was called with patches which were not flipped, it
would return incorrect results, due to the definition of mathematical
convolution [1]. Most of the time, this behavior isn't a problem,
because C2.m is called via a wrapper function, extractC2FromCell.m,
which does the flipping. However, it felt to me that the flipping was
actually done in the wrong place, since it's not needed until the
convolution occurs in windowedPatchDistance.m.  So, I've updated the
'develop' branch of the HMAX-MATLAB repository to perform the flip just
before the convolution, rather than in a separate function. I've left
the 'master' branch unchanged, since it is technically correct. These
changes were tested to give identical results within numerical precision
for randomly generated images.